쿠폰 발급 API 및 보유 쿠폰 목록 조회 API 를 작성

+++ src/main/java/kr/hhplus/be/server/payment/entity/CustomerPointHistory.java
@@ -3,7 +3,10 @@ package kr.hhplus.be.server.payment.entity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import kr.hhplus.be.server.customer.entity.Customers;
 import kr.hhplus.be.server.payment.dto.TransactionType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,9 +25,8 @@ public class CustomerPointHistory {
 	@GeneratedValue(strategy = jakarta.persistence.GenerationType.AUTO)
 	long id;
 	long customerId;
-	long point;
+	int point;
 	TransactionType type;
 	long updateMillis;
-
 
 }
